Conform types to windows spec
In this commit I've used
https://technet.microsoft.com/en-us/library/jj554896(v=wps.630).aspx to figure
out the datatype for every property and munge/set-insync? accordingly.  All
boolean values will be converted to strings and capitalized, all actions are
capitalized, and all strings and file paths are downcased (for now).

I may need to deal with the path separately, but I'll need to do some research
on that.